feat(new service (tasks)): tasks Service Implementation
Implemented a new service for handling tasks for each car in the repair shop. The new service
supports full CRUD and communicates via messages with the Cars Service where the new endpoints are
exposed

BREAKING CHANGE: Refactored the DB relations to match the new functionalites. Chaged ports for the
microservices initialization